0-print_listint.c

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	size_t lenght = 0;
+	int lenght = 0;
 
 	while (h != NULL)
 	{
